refactor(components): update typings and simplify imports
Replaced component-specific prop types with a generic `ReactProps` utility for consistency and clarity. Updated imports across multiple components to align with the new typing structure. Improved maintainability and reduced redundancy in prop type definitions.
