+++ src/components/Divided.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { dividerStyle } from '../styles/divider.style';
-import { DividerProps } from '../interfaces/divider.interface';
+import { DividerInterface } from '../interfaces/divider.interface';
+import { ReactProps } from '../utils/component';
 
 export const Divider = ({
   orientation = 'horizontal',
   className,
   ...restProps
-}: DividerProps) => {
+}: ReactProps<DividerInterface>) => {
   const styles = dividerStyle({ orientation, className });
 
   return <hr className={styles.divider} {...restProps} />;
